[@benja] Refactor symptoms-registry into registro-sintomas
