+++ pages/registro-sintomas.ts
