+++ test/pages/registro-sintomas.test.tsx
@@ -7,13 +7,13 @@ import {
   waitFor,
 } from 'test/testUtils'
 import axios from 'axios'
-import SymptomsRegistry from 'pages/symptoms-registry'
+import SymptomsRegistry from 'pages/registro-sintomas'
 
 jest.mock('axios')
-const mockPush = jest.fn().mockResolvedValue(null)
 
+const mockPush = jest.fn().mockResolvedValue(null)
 const mockQuery = {
-  'pain-level': '0',
+  'nivel-dolor': '0',
 }
 jest.mock('next/router', () => ({
   useRouter: () => ({
@@ -25,10 +25,11 @@ jest.mock('next/router', () => ({
 jest.mock('next-auth/client', () => ({
   useSession: jest.fn().mockReturnValue([{ role: 'tutor' }, false]),
 }))
+
 describe('Home page', () => {
   beforeEach(() => {
-    render(<SymptomsRegistry />)
     jest.clearAllMocks()
+    render(<SymptomsRegistry />)
   })
 
   afterEach(cleanup)
@@ -111,7 +112,7 @@ describe('Home page', () => {
   })
 
   test('should not render painbox when painLevel is invalid', () => {
-    mockQuery['pain-level'] = 'bla'
+    mockQuery['nivel-dolor'] = 'bla'
     render(<SymptomsRegistry />)
 
     expect(screen.queryByText(/sin dolor/i)).not.toBeInTheDocument
